requirements.txt
aiofiles
dnspython
fastapi
passlib[bcrypt]
pydantic[email]
pyjwt
python-dotenv
python-multipart
motor
uvicorn
